fix: change type of user id to maintain consistency with mysql and other functions parts of package

+++ product/types.go
@@ -1,7 +1,7 @@
 package product
 
 type AddableProduct struct {
-	UserId             int      `json:"user_id" binding:"required"`
+	UserId             int64    `json:"user_id" binding:"required"`
 	ProductName        string   `json:"product_name" binding:"required"`
 	ProductDescription string   `json:"product_description" binding:"required"` // text sql larger sizze
 	ProductImages      []string `json:"product_images" binding:"required"`
